+++ test/fw/ptl/lib/pbs_testlib.py
@@ -53,4 +53,3 @@ from ptl.lib.ptl_comm import *
 from ptl.lib.ptl_resourceresv import *
 from ptl.lib.ptl_fairshare import *
 from ptl.lib.ptl_entities import *
-from ptl.lib.ptl_wrappers import *
